requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-cloud-aiplatform==1.92.0
google-cloud-storage==2.19.0
google-genai==1.14.0
google-api-python-client==2.155.0
google-auth==2.36.0
pydantic==2.10.0
